// src/models/QueryKey.ts
export enum QueryKey {
  // Auth
  SIGN_UP_MUTATION = 'sign-up-mutation',
  LOGIN_MUTATION = 'login-mutation',
  REQUEST_LOGIN_OTP_MUTATION = 'request-login-otp-mutation',
  VERIFY_LOGIN_OTP_MUTATION = 'verify-login-otp-mutation',
  REQUEST_FORGOT_PASSWORD_OTP_MUTATION = 'request-forgot-password-otp-mutation',
  VERIFY_FORGOT_PASSWORD_OTP_MUTATION = 'verify-forgot-password-otp-mutation',
  CHECK_ADMIN_QUERY = 'check-admin-query',
  //profile
  PROFILE_QUERY = 'profile-query',
  //product
  PRODUCT_LIST = 'PRODUCT_LIST',
  PRODUCT_DETAIL = 'PRODUCT_DETAIL',
  PRODUCT_HOT_SALE = 'PRODUCT_HOT_SALE',
  PRODUCT_FEATURED = 'PRODUCT_FEATURED',
  PRODUCT_RELATED = 'PRODUCT_RELATED',
  PRODUCT_LIST_BY_CATEGORY = 'PRODUCT_LIST_BY_CATEGORY',
  PRODUCT_LIST_BY_SUB_CATEGORY = 'PRODUCT_LIST_BY_SUB_CATEGORY',
  PRODUCT_LIST_BY_PRODUCT_TYPE = 'PRODUCT_LIST_BY_PRODUCT_TYPE',
  PRODUCT_LIST_BY_BRAND = 'PRODUCT_LIST_BY_BRAND',
  //brand
  BRAND_LIST = 'BRAND_LIST',
  //category
  CATEGORY_LIST = 'CATEGORY_LIST',
  CATEGORY_MENU = 'CATEGORY_MENU',
  //product type
  PRODUCT_TYPE_LIST = 'PRODUCT_TYPE_LIST',
  //sub category
  SUB_CATEGORY_LIST = 'SUB_CATEGORY_LIST',
  SUB_CATEGORY_BY_CATEGORY = 'SUB_CATEGORY_BY_CATEGORY',
  PRODUCT_TYPE_BY_SUB_CATEGORY = 'PRODUCT_TYPE_BY_SUB_CATEGORY',
  PRODUCT_TYPE_DETAIL = 'PRODUCT_TYPE_DETAIL',
  //supplier
  SUPPLIER_LIST = 'SUPPLIER_LIST',
  SUPPLIER_DETAIL = 'SUPPLIER_DETAIL',
  INVOICE_LIST = 'INVOICE_LIST',
  INVOICE_DETAIL = 'INVOICE_DETAIL',
  INVOICE_ITEM_LIST = 'INVOICE_ITEM_LIST',
  AUTH_ME = 'AUTH_ME',

  // Add discount-related keys
  DISCOUNT_LIST = 'DISCOUNT_LIST',
  DISCOUNT_DETAIL = 'DISCOUNT_DETAIL',
  DISCOUNT_BY_CODE = 'DISCOUNT_BY_CODE',

  // User management
  USER_LIST = 'USER_LIST',
  USER_PROFILE = 'USER_PROFILE',

  // Wishlist
  WISHLIST_LIST = 'WISHLIST_LIST',
  MY_WISHLIST = 'MY_WISHLIST',

  // Address
  ADDRESS_LIST = 'ADDRESS_LIST',
  ADDRESS_DETAIL = 'ADDRESS_DETAIL',

  // Cart
  GET_CART = 'GET_CART',

  // Order
  MY_ORDERS = 'MY_ORDERS',
  ORDER_DETAIL = 'ORDER_DETAIL',
  ORDER_LIST = 'ORDER_LIST',

  // Notification
  GET_NOTIFICATIONS_USER = 'GET_NOTIFICATIONS_USER',
}